RenameDesiredNames: move package
